feat: Replace Markaai image with Ann Arbor skyline
The old image was from Pexels and had license problems. In
this commit, we replace the Markaai image, which was originally
an image of Dubai's skyline, with an image of the skyline of
Ann Arbor, Michigan.

In addition to this, we add a high-dpi version of the image
and update the readme's copyright notices.