feat: Add mission to pickup anime from Alexandria
We add a mission given to us by the Anime Organization. To start
the mission, your reputation with the Republic and the Anime
Organization must be 50 or higher and your cargo space must be
at least 140.

This has a 20% chance of triggering everytime you enter the
spaceport of an Anime Organization planet.

And of course, the private collector just had to be a middle-aged
lady wearing a white pantsuit. Her name is unfortunately, not
revealed in the mission. Not yet, that is.