feat: Reduce choices in Debian hat mission
In the mission where you buy someone a Debian hat, you had the option to
indicate what distro you used to the person, and if you informed him
that you used Debian, you would get the mission.

However, there were too many distros to choose from, so in this commit I
remove most of the choices and only allow the player to choose between
Ubuntu, Arch Linux, and Debian, which the person wants.

v0.10.16-patch32